Dom's comments
