+++ geoh5py/ui_json/forms.py
@@ -48,9 +48,6 @@ from geoh5py.shared.validators import (
     types_to_string,
 )
 from geoh5py.ui_json.annotations import OptionalUUIDList, OptionalValueList
-from geoh5py.ui_json.form_utils import (
-    filter_candidates_by_indicator_polling,
-)
 from geoh5py.ui_json.validations.form import (
     empty_string_to_none,
     uuid_to_string,
@@ -122,8 +119,6 @@ class BaseForm(BaseModel):
     def infer(
         cls,
         data: dict[str, Any],
-        form_types: list[type[BaseForm]],
-        indicators: list[set[str]],
     ) -> type[BaseForm]:
         """
         Infer and return the appropriate form.
@@ -146,7 +141,7 @@ class BaseForm(BaseModel):
         data = {to_snake(k): v for k, v in data.items()}
 
         candidates = filter_candidates_by_indicator_polling(
-            indicators, data, form_types
+            INDICATORS, data, FORM_TYPES
         )
         if len(candidates) == 1:
             return candidates[0]
@@ -588,3 +583,71 @@ class DataRangeForm(DataFormMixin, BaseForm):
     range_label: str
     allow_complement: bool = False
     is_complement: bool = False
+
+
+def all_subclasses(type_object: type[BaseForm]) -> list[type[BaseForm]]:
+    """Recursively find all subclasses of input type object."""
+    collection = []
+    subclasses = type_object.__subclasses__()
+    collection += subclasses
+    for subclass in subclasses:
+        collection += all_subclasses(subclass)
+    return collection
+
+
+def model_fields_difference(parent: type[BaseForm], child: type[BaseForm]) -> set[str]:
+    """Isolate fields added in the child class."""
+    return set(child.model_fields) - set(parent.model_fields)
+
+
+def indicator_attributes(
+    parent: type[BaseForm], children: list[type[BaseForm]]
+) -> list[set[str]]:
+    """List all the attributes defined in a subclass."""
+    return [model_fields_difference(parent, c) for c in children]
+
+
+def filter_candidates_by_indicator_polling(
+    indicators: list[set[str]],
+    data: dict[str, Any],
+    candidates: list[type[BaseForm]],
+) -> np.ndarray:
+    """
+    Return candidate subclass(es) with most matching indicators.
+
+    Polling will return a single correct candidate subclass if the
+    form data includes any unique indicators.  It will also resolve
+    any ambiguity between non-unique indicators such as 'choice_list'
+    and 'multi_select'.
+
+    """
+    counts = count_indicators(indicators, data)
+    candidates = np.array(candidates)[counts == np.max(counts)]
+    if len(candidates) > 1 and len(candidates) < len(indicators):
+        candidates = baseclass_if_equal_indicators(candidates)
+
+    return candidates
+
+
+def count_indicators(indicators: list[set[str]], data: dict[str, Any]) -> np.ndarray:
+    """Return the number of matching indicators for each child class."""
+    return np.array([len(i.intersection(data)) for i in indicators])
+
+
+def baseclass_if_equal_indicators(candidates: np.ndarray) -> np.ndarray:
+    """
+    Choose the base class for overlapping indicators.
+
+    If multiple subclasses returned the same number of matching indicators,
+    this is because they form a hierarchy and the correct choice is the base
+    class.  Eg. If 'parent', 'association', and 'data_type' are indicators
+    for a particular form, then all forms will count three matching
+    indicators. However, the correct choice is the form with the least number
+    of attributes, (DataForm).
+    """
+    n_attributes = [len(k.model_fields) for k in candidates]
+    return candidates[n_attributes == np.min(n_attributes)]
+
+
+FORM_TYPES = all_subclasses(BaseForm)
+INDICATORS = indicator_attributes(BaseForm, FORM_TYPES)
