+++ tests/ui_json/forms_test.py
@@ -29,11 +29,6 @@ from pydantic import BaseModel, ValidationError
 from geoh5py import Workspace
 from geoh5py.groups import GroupTypeEnum, PropertyGroup
 from geoh5py.objects import Curve, Points, Surface
-from geoh5py.ui_json.form_utils import (
-    all_subclasses,
-    indicator_attributes,
-    model_fields_difference,
-)
 from geoh5py.ui_json.forms import (
     Association,
     BaseForm,
@@ -53,6 +48,9 @@ from geoh5py.ui_json.forms import (
     ObjectForm,
     RadioLabelForm,
     StringForm,
+    all_subclasses,
+    indicator_attributes,
+    model_fields_difference,
 )
 from geoh5py.ui_json.ui_json import BaseUIJson
 
@@ -609,26 +607,20 @@ def test_flatten(sample_form):
 
 
 def test_base_form_infer(tmp_path):
-    form_types = all_subclasses(BaseForm)
-    indicators = indicator_attributes(BaseForm, form_types)
-    form = BaseForm.infer({"label": "test", "value": "test"}, form_types, indicators)
+    form = BaseForm.infer({"label": "test", "value": "test"})
     assert form == StringForm
-    form = BaseForm.infer({"label": "test", "value": 1}, form_types, indicators)
+    form = BaseForm.infer({"label": "test", "value": 1})
     assert form == IntegerForm
-    form = BaseForm.infer({"label": "test", "value": 1.0}, form_types, indicators)
+    form = BaseForm.infer({"label": "test", "value": 1.0})
     assert form == FloatForm
-    form = BaseForm.infer({"label": "test", "value": True}, form_types, indicators)
+    form = BaseForm.infer({"label": "test", "value": True})
     assert form == BoolForm
     form = BaseForm.infer(
         {"label": "test", "value": str(uuid.uuid4()), "meshType": [Points]},
-        form_types,
-        indicators,
     )
     assert form == ObjectForm
     form = BaseForm.infer(
         {"label": "test", "value": str(uuid.uuid4()), "mesh_type": Points},
-        form_types,
-        indicators,
     )
     assert form == ObjectForm
     form = BaseForm.infer(
@@ -639,8 +631,6 @@ def test_base_form_infer(tmp_path):
             "association": "Vertex",
             "dataType": "Float",
         },
-        form_types,
-        indicators,
     )
     assert form == DataForm
     form = BaseForm.infer(
@@ -652,8 +642,6 @@ def test_base_form_infer(tmp_path):
             "dataType": "Float",
             "multiSelect": True,
         },
-        form_types,
-        indicators,
     )
     assert form == MultiSelectDataForm
     form = BaseForm.infer(
@@ -666,14 +654,10 @@ def test_base_form_infer(tmp_path):
             "isValue": True,
             "property": str(uuid.uuid4()),
         },
-        form_types,
-        indicators,
     )
     assert form == DataOrValueForm
     form = BaseForm.infer(
         {"label": "test", "groupType": PropertyGroup, "value": str(uuid.uuid4())},
-        form_types,
-        indicators,
     )
     assert form == GroupForm
     form = BaseForm.infer(
@@ -684,14 +668,10 @@ def test_base_form_infer(tmp_path):
             "fileType": ["ext"],
             "fileDescription": ["something"],
         },
-        form_types,
-        indicators,
     )
     assert form == FileForm
     form = BaseForm.infer(
         {"label": "test", "value": "test", "choiceList": ["test", "other"]},
-        form_types,
-        indicators,
     )
     assert form == ChoiceForm
     form = BaseForm.infer(
@@ -701,8 +681,6 @@ def test_base_form_infer(tmp_path):
             "value": ["test", "other"],
             "choice_list": ["test", "other", "another"],
         },
-        form_types,
-        indicators,
     )
     assert form == MultiChoiceForm
     form = BaseForm.infer(
@@ -715,8 +693,6 @@ def test_base_form_infer(tmp_path):
             "dataType": "Float",
             "rangeLabel": "value range",
         },
-        form_types,
-        indicators,
     )
     assert form == DataRangeForm
